Harden relay alarm and hardware config state

+++ lib/ssvcOpenConnect/core/StatefulServices/OpenConnectHardwareSettingsService/OpenConnectHardwareConfig.h
@@ -31,7 +31,7 @@ struct OpenConnectHardwareConfig {
     bool lcd1602Enabled = false;
     uint8_t lcd1602I2cAddress = 0x27;
 
-    static bool isValidI2c7Bit(uint8_t addr) { return addr >= 0x08 && addr <= 0x77; }
+    static bool isValidI2c7Bit(unsigned int addr) { return addr >= 0x08 && addr <= 0x77; }
 
     static constexpr size_t kMaxRelayChips = 8;
